Destroy the ship randomly
