+++ src/index.ts
@@ -2,9 +2,9 @@ import { createHud } from "./hud";
 import { createShip } from "./ship";
 import { keyboardControls } from "./keyboardControls";
 import { createGround } from "./ground";
+import { mountExplosion, showExplosion } from "./explosions";
 
 import "./index.css";
-import { svgPoints } from "./ui";
 
 const keyboard = keyboardControls();
 const ship = createShip(keyboard);
@@ -13,8 +13,21 @@ const ground = createGround();
 const hud = createHud({ getShipStats: ship.readState });
 
 let lastFrameTime = Date.now();
+let isGameOver = false;
 
 function update() {
+  // we end the game randomly until we detect collisions
+  const MISFORTUNE = 0.0025;
+  let luck = Math.random();
+  if (luck < MISFORTUNE) {
+    isGameOver = true;
+  }
+
+  if (isGameOver) {
+    showExplosion(ship.readState().position);
+    return;
+  }
+
   const now = Date.now();
   const deltaTime = now - lastFrameTime;
 
@@ -22,6 +35,7 @@ function update() {
   ship.update(deltaTime);
 
   lastFrameTime = now;
+
   requestAnimationFrame(() => update());
 }
 
@@ -29,32 +43,8 @@ document.getElementById("hud")!.innerHTML = hud.render();
 document.getElementById("canvas")!.innerHTML = `
   ${ship.render()}
   ${ground.render()}
-  ${mountExplosion()}
+  ${mountExplosion({ id: "ship-explosion" })}
 `;
 
 update();
 hud.scheduleUpdates();
-
-function mountExplosion() {
-  const outerRadius = 14;
-  const innerRadius = 7;
-  const strokeWidth = 1 + Math.round(Math.random());
-
-  const spikes = Math.floor(Math.random() * 4) + 5; // Random number of spikes between 10 and 15
-  const explosionPoints = Array.from({ length: spikes * 2 }, (_, i) => {
-    const angle = (Math.PI / spikes) * i * (1.05 - Math.random() * 0.1);
-    let radius = i % 2 === 0 ? innerRadius : outerRadius;
-    const x = 5 + radius * Math.cos(angle);
-    const y = 5 + radius * Math.sin(angle);
-    return { x: Math.floor(x), y: Math.floor(y) };
-  });
-  explosionPoints.push(explosionPoints[0]);
-
-  return `
-    <svg x="40" y="40" class="explosion">
-      <polyline stroke-width="${strokeWidth}" points="${svgPoints(
-    explosionPoints
-  )}" />
-    </svg>
-  `;
-}
